Updated metadata management
Instead of storing each metadata file inside the addon folder, each metadata is collected and stored in directory inside %LOCALAPPDATA%\YAAM\metadata

This prevents sharding of the metadata items and cluttering other directories with application specific stuff

To avoid splitting between local and global addons, metadata are placed in bucket where the bucket name is the hash of the addon path. This ensures that application using the same addons are transparently using the same metadata.